feat: Save company and job after saving page

+++ src/main/java/com/jobosint/service/ai/JobDescriptionParserService.java
@@ -1,10 +1,8 @@
 package com.jobosint.service.ai;
 
-import com.jobosint.model.JobDescription;
 import com.jobosint.model.ai.JobDescriptionParseResult;
 import com.jobosint.parse.JobDescriptionParser;
 import com.jobosint.parse.LinkedInParser;
-import com.jobosint.parse.ParseResult;
 import com.jobosint.service.TokenizerService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -16,6 +14,7 @@ import org.springframework.ai.openai.api.OpenAiApi;
 import org.springframework.ai.parser.BeanOutputParser;
 import org.springframework.stereotype.Service;
 
+import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -32,16 +31,20 @@ public class JobDescriptionParserService {
     private final LinkedInParser linkedInParser;
     private final TokenizerService tokenizerService;
 
-    public Optional<JobDescriptionParseResult> parseJobDescription() {
+    public Optional<JobDescriptionParseResult> parseJobDescription(String path) {
         var outputParser = new BeanOutputParser<>(JobDescriptionParseResult.class);
 
-        String filePath = "/home/shane/projects/jobosint/data/pages/20240222-0800/3ffa83f1-1fe1-4e65-9629-ae9599f9fb2a.html";
-
         // first let's pare down the raw content
+        File contentFile = new File(path);
+
+        if (!contentFile.exists()) {
+            log.error("File not found: {}", path);
+            return Optional.empty();
+        }
 
         String jd;
         try {
-            jd = linkedInParser.parseJobDescription(filePath);
+            jd = linkedInParser.parseJobDescription(path);
             Files.writeString(Path.of("jd.txt"), jd);
         } catch (IOException e) {
             throw new RuntimeException(e);
